express calculator
Former-commit-id: 769c5e7eed4eced6636b8adaf3c7b7f3a9df0390

+++ core/domain/interface/order/order.go
@@ -541,7 +541,7 @@ type (
 		// 商店编号
 		ShopId int `db:"-"`
 		// 重量,用于生成订单时存储数据
-		Weight int `db:"-"`
+		Weight float32 `db:"-"`
 		// 快递模板编号
 		ExpressTplId int `db:"-"`
 	}
